CV part is integrated in client infrastructure

+++ protocol/MathDataMessage.go
@@ -2,6 +2,6 @@ package protocol
 
 // MathData contains data from measurements
 type MathData struct {
-	DataArray []float64 `json:"data_array"`
+	DataArray []uint32  `json:"data_array"`
 	TimeArray []float64 `json:"time_array"`
 }
